Move Web.PathPieces content to .Internal module and export only classes
